fixed bug and added bitmap allocator and deallocator

+++ src/bootloader/part2/memory.h
@@ -8,7 +8,7 @@
 
 void* memcpy(void* dst, const void* src, uint16_t num);
 int memcmp(const void* ptr1, const void* ptr2, uint16_t num);
-void* memset( void* ptr1, int c, uint16_t num);
+void* memset( void* ptr1, int c, uint64_t num);
 void Memory_Detect(MemoryInfo* memoryInfo);
 
 #endif
