+++ src/kernel/memory.h
@@ -3,4 +3,4 @@
 
 int memcpy(const void* dst, const void* src, uint16_t num);
 int memcmp(const void* ptr1, const void* ptr2, uint16_t num);
-void* memset( void* ptr1, int c, uint16_t num);
+void* memset(void* ptr1, int c, uint32_t num);
